+++ src/app/app.component.ts
@@ -1,9 +1,16 @@
 import { Component } from '@angular/core';
 
+// Firebase
+import * as firebase from "firebase/app";
+
 @Component({
     selector: 'app-root',
     templateUrl: './app.component.html',
     styleUrls: ['./app.component.scss']
 })
 export class AppComponent {
+    
+    constructor() {
+        
+    }
 }